Replace int factor to float factor in Vec2

+++ SharpEngine/Utils/Vec2.cs
@@ -37,24 +37,24 @@ namespace SharpEngine
         public static Vec2 operator -(Vec2 vec, Vec2 vec2)
             => new Vec2(vec.x - vec2.x, vec.y - vec2.y);
 
-        public static Vec2 operator -(Vec2 vec, int factor)
+        public static Vec2 operator -(Vec2 vec, float factor)
             => new Vec2(vec.x - factor, vec.y - factor);
         public static Vec2 operator +(Vec2 vec, Vec2 vec2)
             => new Vec2(vec.x + vec2.x, vec.y + vec2.y);
 
-        public static Vec2 operator +(Vec2 vec, int factor)
+        public static Vec2 operator +(Vec2 vec, float factor)
             => new Vec2(vec.x + factor, vec.y + factor);
 
         public static Vec2 operator *(Vec2 vec, Vec2 vec2)
             => new Vec2(vec.x * vec2.x, vec.y * vec2.y);
 
-        public static Vec2 operator *(Vec2 vec, int factor)
+        public static Vec2 operator *(Vec2 vec, float factor)
             => new Vec2(vec.x * factor, vec.y * factor);
 
         public static Vec2 operator /(Vec2 vec, Vec2 vec2)
             => new Vec2(vec.x / vec2.x, vec.y / vec2.y);
 
-        public static Vec2 operator /(Vec2 vec, int factor)
+        public static Vec2 operator /(Vec2 vec, float factor)
             => new Vec2(vec.x / factor, vec.y / factor);
 
         public static implicit operator Vec2(Vector2 vec)
